refactor: create Flutter-free entry point for worker threads
Architectural Improvements:
==========================

Split the library into two entry points to enable DPoP token generation
in Dart isolates and web workers without Flutter dependencies:

1. package:solid_auth/solid_auth.dart (Main entry point)
   - Full Flutter-enabled authentication library
   - SolidAuth class with UI integration
   - OIDC flow management with browser redirects
   - Session persistence using platform storage
   - Exports: SolidAuth, UserAndWebId, DpopCredentials, DPoP

2. package:solid_auth/worker.dart (Worker entry point - NEW)
   - Pure Dart, zero Flutter dependencies
   - Minimal API for DPoP token generation
   - Safe for use in isolates and web workers
   - Exports: DpopCredentials, DPoP, KeyPair

File Structure Changes:
======================

Moved/Created:
- lib/worker.dart (NEW) - Flutter-free public API
- lib/src/gen_dpop_token.dart (MOVED from solid_auth_client.dart)
- lib/src/oidc/dpop_credentials.dart (MOVED from solid_oidc_user_manager.dart)
- example/lib/dpop_worker_example.dart (MOVED from example/)

Deleted:
- lib/src/solid_auth_client.dart (split into gen_dpop_token.dart + imports)
- example/dpop_worker_example.dart (moved to example/lib/)

Modified:
- lib/src/oidc/solid_oidc_user_manager.dart
  - Removed DpopCredentials and DPoP classes (moved to dpop_credentials.dart)
  - Updated imports to use gen_dpop_token.dart
- lib/src/solid_auth.dart
  - Updated exports to reflect new file structure
- test/dpop_credentials_test.dart
  - Updated import to use new dpop_credentials.dart location

Benefits:
========

Performance:
- Offload cryptographic operations from main UI thread
- Generate multiple DPoP tokens in parallel
- Better responsiveness for high-throughput scenarios

Architecture:
- Clear separation of concerns (Flutter vs. pure Dart)
- Enables testing without Flutter test harness
- Potential reuse in non-Flutter Dart projects

Developer Experience:
- Explicit import paths prevent accidental Flutter deps in workers
- Comprehensive documentation in worker.dart
- Complete working examples

Documentation:
=============

Added comprehensive documentation covering:
- worker.dart: 90+ lines explaining architecture and usage
- gen_dpop_token.dart: RFC references and internal API docs
- README.md: Complete worker thread section with examples
- example/lib/dpop_worker_example.dart: Full working patterns

Security Model:
==============

Maintains existing security posture:
- Credentials safe for intra-process transfer only
- No persistent storage of serialized credentials
- Fresh token generation per request
- Private keys never leave credential objects

Testing:
=======

All existing tests pass:
- test/dpop_credentials_test.dart: 7/7 tests passing
- No behavioral changes to public API
- Example code analyzes without issues
